src/third_party/blink/web_tests/wpt_internal/scheduler/current_task_signal_inherit_priority.html

<!doctype html>
<title>Scheduling API: Signal inheritance</title>
<link rel="author" title="Nate Chapin" href="mailto:japhet@chromium.org">
<link rel="help" href="https://github.com/WICG/main-thread-scheduling">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
'use strict';

async_test(t => {
  scheduler.postTask(t.step_func_done(() => {
    assert_equals('user-blocking', scheduler.currentTaskSignal.priority);
  }), { priority: "user-blocking" });
}, 'Test that currentTaskSignal propagates priority even if an explicit signal was not given');

</script>
